Move apiUrl to env
