+++ src/environments/environment.ts
@@ -1,0 +1,4 @@
+export const environment = {
+  production: true,
+  apiUrl: 'https://bastet-server-ef94bb4e91eb.herokuapp.com/api',
+};
